moved atomic_structure, lattice and basic_types
