+++ examples/json_structure.cc
@@ -26,7 +26,6 @@
  * Boston, MA 02111-1307, USA.
  */
 
-#include "rascal/basic_types.hh"
 #include "rascal/representations/calculator_sorted_coulomb.hh"
 #include "rascal/representations/calculator_spherical_expansion.hh"
 #include "rascal/representations/calculator_spherical_invariants.hh"
@@ -34,6 +33,7 @@
 #include "rascal/structure_managers/adaptor_strict.hh"
 #include "rascal/structure_managers/make_structure_manager.hh"
 #include "rascal/structure_managers/structure_manager_centers.hh"
+#include "rascal/utils/basic_types.hh"
 #include "rascal/utils/json_io.hh"
 #include "rascal/utils/utils.hh"
