+++ src/rascal/structure_managers/adaptor_increase_maxorder.hh
@@ -30,10 +30,10 @@
 #ifndef SRC_RASCAL_STRUCTURE_MANAGERS_ADAPTOR_INCREASE_MAXORDER_HH_
 #define SRC_RASCAL_STRUCTURE_MANAGERS_ADAPTOR_INCREASE_MAXORDER_HH_
 
-#include "rascal/basic_types.hh"
-#include "rascal/lattice.hh"
+#include "rascal/structure_managers/lattice.hh"
 #include "rascal/structure_managers/property.hh"
 #include "rascal/structure_managers/structure_manager.hh"
+#include "rascal/utils/basic_types.hh"
 #include "rascal/utils/utils.hh"
 
 #include <algorithm>
