+++ src/rascal/structure_managers/structure_manager_centers.hh
@@ -34,10 +34,10 @@
 
 // inclusion of librascal data structure, each manager is based on the interface
 // given in `structure_manager.hh`
-#include "rascal/atomic_structure.hh"
-#include "rascal/basic_types.hh"
-#include "rascal/lattice.hh"
+#include "rascal/structure_managers/atomic_structure.hh"
+#include "rascal/structure_managers/lattice.hh"
 #include "rascal/structure_managers/structure_manager.hh"
+#include "rascal/utils/basic_types.hh"
 #include "rascal/utils/json_io.hh"
 
 // data types and operations are based on the Eigen library
